Module aws_sdk_polly::error

Modules

engine_not_supported_exception

See EngineNotSupportedException

invalid_lexicon_exception

See InvalidLexiconException

invalid_next_token_exception

See InvalidNextTokenException

invalid_s3_bucket_exception

See InvalidS3BucketException

invalid_s3_key_exception

See InvalidS3KeyException

invalid_sample_rate_exception

See InvalidSampleRateException

invalid_sns_topic_arn_exception

See InvalidSnsTopicArnException

invalid_ssml_exception

See InvalidSsmlException

invalid_task_id_exception

See InvalidTaskIdException

language_not_supported_exception

See LanguageNotSupportedException

lexicon_not_found_exception

See LexiconNotFoundException

lexicon_size_exceeded_exception

See LexiconSizeExceededException

marks_not_supported_for_format_exception

See MarksNotSupportedForFormatException

max_lexeme_length_exceeded_exception

See MaxLexemeLengthExceededException

max_lexicons_number_exceeded_exception

See MaxLexiconsNumberExceededException

service_failure_exception

See ServiceFailureException

ssml_marks_not_supported_for_text_type_exception

See SsmlMarksNotSupportedForTextTypeException

synthesis_task_not_found_exception

See SynthesisTaskNotFoundException

text_length_exceeded_exception

See TextLengthExceededException

unsupported_pls_alphabet_exception

See UnsupportedPlsAlphabetException

unsupported_pls_language_exception

See UnsupportedPlsLanguageException

Structs

DeleteLexiconError

DescribeVoicesError

EngineNotSupportedException

This engine is not compatible with the voice that you have designated. Choose a new voice

GetLexiconError

GetSpeechSynthesisTaskError

InvalidLexiconException

Amazon Polly can't find the specified lexicon. Verify that the lexicon's name is

InvalidNextTokenException

The NextToken is invalid. Verify that it's spelled correctly, and then try

InvalidS3BucketException

The provided Amazon S3 bucket name is invalid. Please check your input with S3 bucket

InvalidS3KeyException

The provided Amazon S3 key prefix is invalid. Please provide a valid S3 object key

InvalidSampleRateException

The specified sample rate is not valid.

InvalidSnsTopicArnException

The provided SNS topic ARN is invalid. Please provide a valid SNS topic ARN and try

InvalidSsmlException

The SSML you provided is invalid. Verify the SSML syntax, spelling of tags and values,

InvalidTaskIdException

The provided Task ID is not valid. Please provide a valid Task ID and try again.

LanguageNotSupportedException

The language specified is not currently supported by Amazon Polly in this capacity.

LexiconNotFoundException

Amazon Polly can't find the specified lexicon. This could be caused by a lexicon that

LexiconSizeExceededException

The maximum size of the specified lexicon would be exceeded by this

ListLexiconsError

ListSpeechSynthesisTasksError

MarksNotSupportedForFormatException

Speech marks are not supported for the OutputFormat selected. Speech marks

MaxLexemeLengthExceededException

The maximum size of the lexeme would be exceeded by this operation.

MaxLexiconsNumberExceededException

The maximum number of lexicons would be exceeded by this operation.

PutLexiconError

ServiceFailureException

An unknown condition has caused a service failure.

SsmlMarksNotSupportedForTextTypeException

SSML speech marks are not supported for plain text-type input.

StartSpeechSynthesisTaskError

SynthesisTaskNotFoundException

The Speech Synthesis task with requested Task ID cannot be found.

SynthesizeSpeechError

TextLengthExceededException

The value of the "Text" parameter is longer than the accepted limits. For the

UnsupportedPlsAlphabetException

The alphabet specified by the lexicon is not a supported alphabet. Valid values are

UnsupportedPlsLanguageException

The language specified in the lexicon is unsupported. For a list of supported

Enums

DeleteLexiconErrorKind

DescribeVoicesErrorKind

GetLexiconErrorKind

GetSpeechSynthesisTaskErrorKind

ListLexiconsErrorKind

ListSpeechSynthesisTasksErrorKind

PutLexiconErrorKind

StartSpeechSynthesisTaskErrorKind

SynthesizeSpeechErrorKind