Module aws_sdk_polly::model

Modules

lexicon

See Lexicon

lexicon_attributes

See LexiconAttributes

lexicon_description

See LexiconDescription

synthesis_task

See SynthesisTask

voice

See Voice

Structs

Lexicon

Provides lexicon name and lexicon content in string format. For more information, see

LexiconAttributes

Contains metadata describing the lexicon such as the number of lexemes, language code,

LexiconDescription

Describes the content of the lexicon.

SynthesisTask

SynthesisTask object that provides information about a speech synthesis task.

Voice

Description of the voice.

Enums

Engine

Gender

LanguageCode

OutputFormat

SpeechMarkType

TaskStatus

TextType

VoiceId