Struct aws_sdk_polly::client::fluent_builders::StartSpeechSynthesisTask

pub struct StartSpeechSynthesisTask { /* fields omitted */ }

Implementations

impl StartSpeechSynthesisTask

pub async fn send(
self
) -> Result<StartSpeechSynthesisTaskOutput, SdkError<StartSpeechSynthesisTaskError>>

pub fn engine(self, inp: Engine) -> Self

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

pub fn set_engine(self, inp: Option<Engine>) -> Self

pub fn language_code(self, inp: LanguageCode) -> Self

Optional language code for the Speech Synthesis request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly will use the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

pub fn set_language_code(self, inp: Option<LanguageCode>) -> Self

pub fn lexicon_names(self, inp: impl Into<String>) -> Self

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

pub fn set_lexicon_names(self, inp: Option<Vec<String>>) -> Self

pub fn output_format(self, inp: OutputFormat) -> Self

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

pub fn set_output_format(self, inp: Option<OutputFormat>) -> Self

pub fn output_s3_bucket_name(self, inp: impl Into<String>) -> Self

Amazon S3 bucket name to which the output file will be saved.

pub fn set_output_s3_bucket_name(self, inp: Option<String>) -> Self

pub fn output_s3_key_prefix(self, inp: impl Into<String>) -> Self

The Amazon S3 key prefix for the output speech file.

pub fn set_output_s3_key_prefix(self, inp: Option<String>) -> Self

pub fn sample_rate(self, inp: impl Into<String>) -> Self

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

pub fn set_sample_rate(self, inp: Option<String>) -> Self

pub fn sns_topic_arn(self, inp: impl Into<String>) -> Self

ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

pub fn set_sns_topic_arn(self, inp: Option<String>) -> Self

pub fn speech_mark_types(self, inp: impl Into<SpeechMarkType>) -> Self

The type of speech marks returned for the input text.

pub fn set_speech_mark_types(self, inp: Option<Vec<SpeechMarkType>>) -> Self

pub fn text(self, inp: impl Into<String>) -> Self

The input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

pub fn set_text(self, inp: Option<String>) -> Self

pub fn text_type(self, inp: TextType) -> Self

Specifies whether the input text is plain text or SSML. The default value is plain text.

pub fn set_text_type(self, inp: Option<TextType>) -> Self

pub fn voice_id(self, inp: VoiceId) -> Self

Voice ID to use for the synthesis.

pub fn set_voice_id(self, inp: Option<VoiceId>) -> Self