Module aws_sdk_polly::error

Modules

engine_not_supported_error

See EngineNotSupportedError

invalid_lexicon_error

See InvalidLexiconError

invalid_next_token_error

See InvalidNextTokenError

invalid_s3_bucket_error

See InvalidS3BucketError

invalid_s3_key_error

See InvalidS3KeyError

invalid_sample_rate_error

See InvalidSampleRateError

invalid_sns_topic_arn_error

See InvalidSnsTopicArnError

invalid_ssml_error

See InvalidSsmlError

invalid_task_id_error

See InvalidTaskIdError

language_not_supported_error

See LanguageNotSupportedError

lexicon_not_found_error

See LexiconNotFoundError

lexicon_size_exceeded_error

See LexiconSizeExceededError

marks_not_supported_for_format_error

See MarksNotSupportedForFormatError

max_lexeme_length_exceeded_error

See MaxLexemeLengthExceededError

max_lexicons_number_exceeded_error

See MaxLexiconsNumberExceededError

service_failure_error

See ServiceFailureError

ssml_marks_not_supported_for_text_type_error

See SsmlMarksNotSupportedForTextTypeError

synthesis_task_not_found_error

See SynthesisTaskNotFoundError

text_length_exceeded_error

See TextLengthExceededError

unsupported_pls_alphabet_error

See UnsupportedPlsAlphabetError

unsupported_pls_language_error

See UnsupportedPlsLanguageError

Structs

DeleteLexiconError

DescribeVoicesError

EngineNotSupportedError

This engine is not compatible with the voice that you have designated. Choose a new voice

GetLexiconError

GetSpeechSynthesisTaskError

InvalidLexiconError

Amazon Polly can't find the specified lexicon. Verify that the lexicon's name is

InvalidNextTokenError

The NextToken is invalid. Verify that it's spelled correctly, and then try

InvalidS3BucketError

The provided Amazon S3 bucket name is invalid. Please check your input with S3 bucket

InvalidS3KeyError

The provided Amazon S3 key prefix is invalid. Please provide a valid S3 object key

InvalidSampleRateError

The specified sample rate is not valid.

InvalidSnsTopicArnError

The provided SNS topic ARN is invalid. Please provide a valid SNS topic ARN and try

InvalidSsmlError

The SSML you provided is invalid. Verify the SSML syntax, spelling of tags and values,

InvalidTaskIdError

The provided Task ID is not valid. Please provide a valid Task ID and try again.

LanguageNotSupportedError

The language specified is not currently supported by Amazon Polly in this capacity.

LexiconNotFoundError

Amazon Polly can't find the specified lexicon. This could be caused by a lexicon that

LexiconSizeExceededError

The maximum size of the specified lexicon would be exceeded by this

ListLexiconsError

ListSpeechSynthesisTasksError

MarksNotSupportedForFormatError

Speech marks are not supported for the OutputFormat selected. Speech marks

MaxLexemeLengthExceededError

The maximum size of the lexeme would be exceeded by this operation.

MaxLexiconsNumberExceededError

The maximum number of lexicons would be exceeded by this operation.

PutLexiconError

ServiceFailureError

An unknown condition has caused a service failure.

SsmlMarksNotSupportedForTextTypeError

SSML speech marks are not supported for plain text-type input.

StartSpeechSynthesisTaskError

SynthesisTaskNotFoundError

The Speech Synthesis task with requested Task ID cannot be found.

SynthesizeSpeechError

TextLengthExceededError

The value of the "Text" parameter is longer than the accepted limits. For the

UnsupportedPlsAlphabetError

The alphabet specified by the lexicon is not a supported alphabet. Valid values are

UnsupportedPlsLanguageError

The language specified in the lexicon is unsupported. For a list of supported

Enums

DeleteLexiconErrorKind

DescribeVoicesErrorKind

GetLexiconErrorKind

GetSpeechSynthesisTaskErrorKind

ListLexiconsErrorKind

ListSpeechSynthesisTasksErrorKind

PutLexiconErrorKind

StartSpeechSynthesisTaskErrorKind

SynthesizeSpeechErrorKind